Submodule Cleanup - Organized Modals
- Moved all modal-related files into a single 'modal' folder for better organization.
- Updated imports to reflect the new folder structure and resolved any import errors.
