+++ js/template/main.js
@@ -4,9 +4,9 @@ import AbstractTemplate from "./AbstractTemplate.js"; // Import AbstractTemplate
 // Importing components for the application
 import Footer from "./footer.js"; // Footer component for the application footer
 import Navigations from "./navigations.js"; // Navigations component for the application navigation
-import ArcsModal from "./arcsModal.js"; // ArcsModal component for modal functionality
-import LoginModal from "./loginModal.js";
-import SignupModal from "./signupModal.js";
+import ArcsModal from "./modal/arcsModal.js"; // ArcsModal component for modal functionality
+import LoginModal from "./modal/loginModal.js";
+import SignupModal from "./modal/signupModal.js";
 
 /**
  * MainTemplate class extends AbstractTemplate to define the main structure and functionality of the application.
