+++ js/template/modal/loginModal/htmlGenerator.js
