+++ js/template/modal/signupModal/htmlGenerator.js
